Genre Count Bar Chart
Currently outputs a bar chart with the correct x axis values (Genres), but the y axis is all scrambled. The y axis contains all the numbers it should as in the actual counts of each genre, BUT the numbers do not appear in order, I am going to start working on creating frequency bins to see if that solves it.